+++ src/components/medium/index.vue
@@ -1,0 +1,9 @@
+<script>
+export const MediumComponent = {
+  name: 'MediumComponent',
+};
+
+export { default as StorageDetail } from './StorageDetail.vue';
+
+export default MediumComponent
+</script>