+++ src/components/small/index.vue
@@ -1,0 +1,9 @@
+<script>
+export const SmallComponent = {
+  name: 'SmallComponent',
+};
+
+export { default as ModalComponent } from './ModalComponent.vue';
+
+export default SmallComponent
+</script>