+++ src/stores/story.js
@@ -1,23 +1,41 @@
-import { defineStore } from 'pinia'
+import { defineStore } from "pinia";
 
 export const storyData = defineStore({
-  id: 'story',
+  id: "story",
   state: () => ({
     counter: 0,
     storyId: 0,
     story: [],
-    playState: 'play',
+    group: [],
+    groupState: "",
+    playState: "play",
     isWrap: false,
     askInstall: true,
+    innerHeight: 0,
     playedStory: [],
+    theme: "business",
+    themeList: [
+      "garden",
+      "business",
+      "wireframe",
+      "black",
+      "valentine",
+      "retro",
+      "fantasy",
+      "pastel",
+      "autumn",
+      "winter",
+    ],
+    maxStorage: 0,
+    currentStorage: 0,
   }),
   getters: {
     doubleCount: (state) => state.counter * 2,
   },
   actions: {
     increment() {
-      this.counter++
+      this.counter++;
     },
   },
   persist: true,
-})
+});