+++ src/views/GlobalSettingStory.vue
@@ -5,8 +5,10 @@ import { ref } from 'vue';
 const story = storyData()
 const route = useRoute()
 const mode = ref('setting')
-const theStory = route.params
-const storyNames = ref(story.story[theStory.id].title)
+const theIfid = route.params.id
+const findIfid = (element) => element.ifid == theIfid;
+const theStory = story.story.findIndex(findIfid)
+const storyNames = ref(story.story[theStory].title)
 const trueStoryNames = storyNames.value
 const dataName = ref(story.story)
 const listStoryFormats = ref([
@@ -56,7 +58,7 @@ function setName(name) {
     }
   })
   if (hasilnya == 0) {
-    story.story[theStory.id].title = name
+    story.story[theStory].title = name
   }
 }
 
@@ -72,9 +74,9 @@ function download(filename, text) {
 </script>
 
 <template>
-  <div class="h-screen flex flex-col">
+  <div class="h-screen flex flex-col" :style='{height: story.innerHeight + "px"}'>
 
-    <div class="navbar bg-primary shadow-lg">
+    <div class="navbar bg-primary shadow-lg text-primary-content">
       <button tabindex="0" class="btn btn-ghost btn-square" @click='$router.back()'>
         <svg fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="h-5 w-5 self-center m-2">
           <path stroke-linecap="round" stroke-linejoin="round" d="M19.5 12h-15m0 0l6.75 6.75M4.5 12l6.75-6.75" />
@@ -102,46 +104,46 @@ function download(filename, text) {
         <label class="label">
           <span class="label-text">Story Formats</span>
         </label>
-        <select v-model="story.story[theStory.id].storyformats" class="select select-bordered w-full">
+        <select v-model="story.story[theStory].storyformats" class="select select-bordered w-full">
           <option v-for="data, index in listStoryFormats" :value="data.val" :key='index'>{{data.name}}</option>
         </select>
       </div>
       <div class='p-2'>
-        <div v-if="story.story[theStory.id].storyformats == 'chapbook-1'">
+        <div v-if="story.story[theStory].storyformats == 'chapbook-1'">
           A Twine story format emphasizing ease of authoring, multimedia, and playability on many different types of devices. Visit the <a class="text-primary underline" href="https://klembot.github.io/chapbook/guide/" target="_blank" rel="noopener noreferrer">guide</a> for more information.
         </div>
-        <div v-if="story.story[theStory.id].storyformats == 'harlowe-3'">
+        <div v-if="story.story[theStory].storyformats == 'harlowe-3'">
           The default story format for Twine 2. Consult its <a class="text-primary underline" href="https://twine2.neocities.org/" target="_blank" rel="noopener noreferrer">documentation</a>.<br>
           <br>
           License: Zlib
         </div>
-        <div v-if="story.story[theStory.id].storyformats == 'harlowe-2'">
+        <div v-if="story.story[theStory].storyformats == 'harlowe-2'">
           The default story format for Twine 2. See its <a class="text-primary underline" href="https://twine2.neocities.org/" target="_blank" rel="noopener noreferrer">documentation</a>.<br>
           <br>
           License: Zlib
         </div>
-        <div v-if="story.story[theStory.id].storyformats == 'harlowe-1'">
+        <div v-if="story.story[theStory].storyformats == 'harlowe-1'">
           The default story format for Twine 2. See its <a class="text-primary underline" href="https://twine2.neocities.org/" target="_blank" rel="noopener noreferrer">documentation</a>.<br>
           <br>
           License: Zlib
         </div>
-        <div v-if="story.story[theStory.id].storyformats == 'snowman-2'">
+        <div v-if="story.story[theStory].storyformats == 'snowman-2'">
           A minimal story format for authors more experienced with HTML, CSS, and JavaScript. See its <a class="text-primary underline" href="https://videlais.github.io/snowman/2/" target="_blank" rel="noopener noreferrer">documentation</a>.
         </div>
-        <div v-if="story.story[theStory.id].storyformats == 'snowman-1'">
+        <div v-if="story.story[theStory].storyformats == 'snowman-1'">
           A minimal story format for authors more experienced with HTML, CSS, and JavaScript. See its <a class="text-primary underline" href="https://videlais.github.io/snowman/1/" target="_blank" rel="noopener noreferrer">documentation</a>.
         </div>
-        <div v-if="story.story[theStory.id].storyformats == 'sugarcube-2'">
+        <div v-if="story.story[theStory].storyformats == 'sugarcube-2'">
           A full featured, highly customizable story format. See its <a class="text-primary underline" href="http://www.motoslave.net/sugarcube/2/#documentation" target="_blank" rel="noopener noreferrer">documentation</a>.<br>
           <br>
           License: BSD-2-Clause
         </div>
-        <div v-if="story.story[theStory.id].storyformats == 'sugarcube-1'">
+        <div v-if="story.story[theStory].storyformats == 'sugarcube-1'">
           A Twine 2 port of the Twine 1 story format by the same name. See its <a class="text-primary underline" href="http://www.motoslave.net/sugarcube/1/#documentation" target="_blank" rel="noopener noreferrer">documentation</a>.<br>
           <br>
           License: Simplified BSD License
         </div>
-        <div v-if="!story.story[theStory.id].storyformats">
+        <div v-if="!story.story[theStory].storyformats">
           Choose Story Formats. To See It's Descriptions And Link To It's Documentations.
         </div>
       </div>
@@ -150,17 +152,17 @@ function download(filename, text) {
         <label class="label">
           <span class="label-text">Story Will Start At : </span>
         </label>
-        <select  v-model="story.story[theStory.id].startNode" class="select select-bordered w-full">
-          <option v-for="data, index in story.story[theStory.id].passage" :value="data.pid" :key='index'>{{data.name}}</option>
+        <select  v-model="story.story[theStory].startNode" class="select select-bordered w-full">
+          <option v-for="data, index in story.story[theStory].passage" :value="data.pid" :key='index'>{{data.name}}</option>
         </select>
       </div>
       <div class="form-control w-full">
         <label class="label">
           <span class="label-text">IFID</span>
         </label>
-        <input v-model="story.story[theStory.id].ifid" disabled type="text" class="input input-disabled input-bordered w-full" />
+        <input v-model="story.story[theStory].ifid" disabled type="text" class="input input-disabled input-bordered w-full" />
       </div>
-      <button @click="download(story.story[theStory.id].title+'.tweezeldata', JSON.stringify(story.story[theStory.id]))" class="btn btn-primary btn-block my-3">Backup This Story.</button>
+      <button @click="download(story.story[theStory].title+'.tweezeldata', JSON.stringify(story.story[theStory]))" class="btn btn-primary btn-block my-3">Backup This Story.</button>
     </div>
 
     <div class="p-1 outline-none w-full grow flex flex-col" v-if="mode == 'css'">
@@ -170,7 +172,7 @@ function download(filename, text) {
           <span class="label-text">Wrap</span> 
         </label>
       </div>
-      <textarea type="textarea" v-model="story.story[theStory.id].userStyle" :class="[story.isWrap ? 'whitespace-pre' : '', 'text-xs font-mono textarea textarea-bordered resize-none grow w-full p-1']"></textarea>
+      <textarea type="textarea" v-model="story.story[theStory].userStyle" :class="[story.isWrap ? 'whitespace-pre' : '', 'text-xs font-mono textarea textarea-bordered resize-none grow w-full p-1']"></textarea>
     </div>
     
     <div class="p-1 outline-none w-full grow flex flex-col" v-if="mode == 'js'">
@@ -180,7 +182,7 @@ function download(filename, text) {
           <span class="label-text">Wrap</span> 
         </label>
       </div>
-      <textarea type="textarea" v-model="story.story[theStory.id].userScript" :class="[story.isWrap ? 'whitespace-pre' : '', 'text-xs font-mono textarea textarea-bordered resize-none grow w-full p-1']"></textarea>
+      <textarea type="textarea" v-model="story.story[theStory].userScript" :class="[story.isWrap ? 'whitespace-pre' : '', 'text-xs font-mono textarea textarea-bordered resize-none grow w-full p-1']"></textarea>
     </div>
   </div>
 </template>