+++ src/views/StoryEdit.vue
@@ -5,10 +5,16 @@ import { ref } from 'vue';
 import { createToaster } from "@meforma/vue-toaster";
 const story = storyData()
 const route = useRoute()
-const theStory = route.params
-const passageNames = ref(story.story[theStory.id].passage[theStory.pid].name)
+const theIfid = route.params.id
+const findIfid = (element) => element.ifid == theIfid;
+const storyIndex = story.story.findIndex(findIfid)
+const thePid = route.params.pid
+const findPid = (element) => element.pid == thePid;
+const passageIndex = story.story[storyIndex].passage.findIndex(findPid)
+const passageNames = ref(story.story[storyIndex].passage[passageIndex].name)
 const truePassageNames = ref(passageNames.value)
-const dataName = ref(story.story[theStory.id].passage)
+const dataName = ref(story.story[storyIndex].passage)
+
 const toaster = createToaster({
   position : 'top',
   duration : 2000,
@@ -27,7 +33,7 @@ function setName(name) {
     }
   })
   if (hasilnya == 0) {
-    story.story[theStory.id].passage[theStory.pid].name = name
+    story.story[storyIndex].passage[passageIndex].name = name
     toaster.show(`Success Changing Passage Name...`, {type : 'success'})
     passageNames.value = name
     truePassageNames.value = name
@@ -35,13 +41,12 @@ function setName(name) {
     toaster.show(`Passage Name Is Alreadys Exist Or Empty...`, {type : 'error'})
   }
 }
-
 </script>
 
 <template>
-  <div class="h-screen flex flex-col">
+  <div class="h-screen flex flex-col" :style='{height: story.innerHeight + "px"}'>
 
-    <div class="navbar bg-primary shadow-lg">
+    <div class="navbar bg-primary shadow-lg text-primary-content">
       <button tabindex="0" class="btn btn-ghost btn-square" @click='$router.back()'>
         <svg fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="h-5 w-5 self-center m-2">
           <path stroke-linecap="round" stroke-linejoin="round" d="M19.5 12h-15m0 0l6.75 6.75M4.5 12l6.75-6.75" />
@@ -64,13 +69,16 @@ function setName(name) {
           </span>
         </label>
       </div>
-      <div class="form-control">
-        <label class="label cursor-pointer justify-start gap-2">
-          <input @click="story.isWrap = !story.isWrap" type="checkbox" class="toggle toggle-primary" :checked='!story.isWrap' />
-          <span class="label-text">Wrap</span> 
-        </label>
+      <div class="flex flex-row w-full items-center gap-2">
+        <div class="form-control w-fit">
+          <label class="label cursor-pointer justify-start gap-2">
+            <input @click="story.isWrap = !story.isWrap" type="checkbox" class="toggle toggle-primary" :checked='!story.isWrap' />
+            <span class="label-text">Wrap</span> 
+          </label>
+        </div>
       </div>
-      <textarea type="textarea" v-model="story.story[theStory.id].passage[theStory.pid].data" :class="[story.isWrap ? 'whitespace-pre' : '', 'text-xs font-mono textarea textarea-bordered resize-none grow w-full h-full px-1']"></textarea>
+      <img id="output"/>
+      <textarea type="textarea" v-model="story.story[storyIndex].passage[passageIndex].data" :class="[story.isWrap ? 'whitespace-pre' : '', 'text-xs font-mono textarea textarea-bordered resize-none grow w-full h-full px-1']"></textarea>
     </div>
 
   </div>